+++ examples/data-sources/meraki_appliance_traffic_shaping/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_traffic_shaping" "example" {
+  network_id = "L_123456"
+}