+++ examples/resources/meraki_appliance_traffic_shaping/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_traffic_shaping.example "<network_id>"