+++ examples/resources/meraki_appliance_traffic_shaping/resource.tf
@@ -1,0 +1,5 @@
+resource "meraki_appliance_traffic_shaping" "example" {
+  network_id                  = "L_123456"
+  global_bandwidth_limit_down = 5120
+  global_bandwidth_limit_up   = 2048
+}